+++ package/zerobias/example-angular-v2/src/app/pages/boards/create-board-form.ts
@@ -70,7 +70,7 @@ interface CreateForm {
       </div>
 
       <div class="form-code">
-        <app-call-reveal [call]="call()" [response]="exampleResponse" />
+        <app-call-reveal [call]="call()" [response]="exampleResponse" [responseType]="'BoardExtended'" />
       </div>
     </form>
   `,