+++ package/zerobias/example-angular-v2/src/app/pages/projects/edit-project-form.ts
@@ -91,7 +91,7 @@ export interface EditProjectBaseline {
       </div>
 
       <div class="form-code">
-        <app-call-reveal [call]="call()" [response]="exampleResponse" />
+        <app-call-reveal [call]="call()" [response]="exampleResponse" [responseType]="'ProjectExtended'" />
       </div>
     </form>
   `,